+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/alter_constraint_columns_order_new.sql
@@ -1,0 +1,7 @@
+CREATE TABLE public.t1 (
+    a circle,
+    b circle
+);
+
+ALTER TABLE public.t1
+	ADD CONSTRAINT c1 EXCLUDE USING gist (b WITH &&, a WITH &&);