+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/alter_index_columns_order_new.sql
@@ -1,0 +1,7 @@
+CREATE TABLE public.t1 (
+    c1 integer,
+    c2 integer,
+    c3 integer
+);
+
+CREATE INDEX i1 ON public.t1 (c1, c3, c2);