+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/schema/IConstraint.java
@@ -23,5 +23,6 @@ public interface IConstraint extends ISearchPath {
     }
     String getDefinition();
     Set<String> getColumns();
+    boolean containsColumn(String name);
     String getTableName();
 }